Fix Scenario 10.
