+++ swift-combine/Sources/EasyRacer/Publishers+Repeating.swift
@@ -1,8 +1,52 @@
-//
-//  File.swift
-//  
-//
-//  Created by Jack Leow on 7/25/24.
-//
+import Atomics
+import Combine
 
-import Foundation
+extension Publishers {
+    struct Repeating<Output>: Publisher {
+        typealias Failure = Never
+        
+        private let element: Output
+        
+        init(_ element: Output) {
+            self.element = element
+        }
+        
+        func receive<S>(subscriber: S) where S : Subscriber, Never == S.Failure, Output == S.Input {
+            let subscription: some Subscription = RepeatingSubscription(element, subscriber)
+            subscriber.receive(subscription: subscription)
+        }
+        
+        final class RepeatingSubscription<Downstream: Subscriber>: Subscription where Downstream.Input == Output, Downstream.Failure == Never {
+            private let element: Output
+            private let subscriber: Downstream
+            private var active: ManagedAtomic<Bool> = ManagedAtomic(true)
+            
+            init(_ element: Output, _ subscriber: Downstream) {
+                self.element = element
+                self.subscriber = subscriber
+            }
+            
+            func request(_ demand: Subscribers.Demand) {
+                if active.load(ordering: .relaxed) {
+                    Task {
+                        if let count: Int = demand.max {
+                            let nextDemand: Subscribers.Demand? = (0..<count)
+                                .map { _ in subscriber.receive(element) }
+                                .last
+                            if let nextDemand: Subscribers.Demand = nextDemand {
+                                request(nextDemand)
+                            }
+                        } else {
+                            let nextDemand: Subscribers.Demand = subscriber.receive(element)
+                            request(nextDemand.max ?? 0 > 0 ? nextDemand : demand)
+                        }
+                    }
+                }
+            }
+            
+            func cancel() {
+                active.store(false, ordering: .relaxed)
+            }
+        }
+    }
+}
